+++ static/Changelog.md
@@ -6,3 +6,16 @@
 - made booting sequence
 - started Desktop enviroment
 - open Apps
+
+## 31.11.2025
+- started doing the header
+
+
+## TO DO
+- popup when you click on header
+    - layout: https://cdn.mos.cms.futurecdn.net/2L4hdz87QauxcXbXqJJx4H.jpg
+    - song: such a funny way
+- make every site work
+- add desktop icons
+- make it more interactive
+- animations